feat: implement StreamDetailsList component and integrate into FileDetailPage and LibraryDetailPage tooltips style: adjust tooltip styling and layout for better visibility
