+++ frontend/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/i18n.ts","./src/main.tsx","./src/vite-env.d.ts","./src/components/animatedsearchicon.tsx","./src/components/appshell.test.tsx","./src/components/appshell.tsx","./src/components/asyncpanel.tsx","./src/components/distributionlist.tsx","./src/components/loaderpinwheelicon.tsx","./src/components/pathbrowser.tsx","./src/components/pathsegmenttrail.tsx","./src/components/statcard.tsx","./src/components/tooltiptrigger.test.tsx","./src/components/tooltiptrigger.tsx","./src/lib/api.test.ts","./src/lib/api.ts","./src/lib/app-data.tsx","./src/lib/app-version.ts","./src/lib/desktop.ts","./src/lib/format.test.ts","./src/lib/format.ts","./src/lib/hdr.test.ts","./src/lib/hdr.ts","./src/lib/ignore-pattern-sections.ts","./src/lib/library-file-column-widths.ts","./src/lib/library-file-search.ts","./src/lib/library-statistics-settings.test.ts","./src/lib/library-statistics-settings.ts","./src/lib/page-visibility.ts","./src/lib/paginated-files.test.ts","./src/lib/paginated-files.ts","./src/lib/path-display.test.ts","./src/lib/path-display.ts","./src/lib/resolution-categories.ts","./src/lib/scan-jobs.test.tsx","./src/lib/scan-jobs.tsx","./src/lib/settings-panel-state.ts","./src/lib/theme.tsx","./src/pages/dashboardpage.tsx","./src/pages/filedetailpage.test.tsx","./src/pages/filedetailpage.tsx","./src/pages/librariespage.test.tsx","./src/pages/librariespage.tsx","./src/pages/librarydetailpage.test.tsx","./src/pages/librarydetailpage.tsx","./src/test/setup.ts"],"version":"5.9.3"}
+{"root":["./src/app.tsx","./src/i18n.ts","./src/main.tsx","./src/vite-env.d.ts","./src/components/animatedsearchicon.tsx","./src/components/appshell.test.tsx","./src/components/appshell.tsx","./src/components/asyncpanel.tsx","./src/components/distributionlist.tsx","./src/components/loaderpinwheelicon.tsx","./src/components/pathbrowser.tsx","./src/components/pathsegmenttrail.tsx","./src/components/statcard.tsx","./src/components/streamdetailslist.tsx","./src/components/tooltiptrigger.test.tsx","./src/components/tooltiptrigger.tsx","./src/lib/api.test.ts","./src/lib/api.ts","./src/lib/app-data.tsx","./src/lib/app-version.ts","./src/lib/desktop.ts","./src/lib/format.test.ts","./src/lib/format.ts","./src/lib/hdr.test.ts","./src/lib/hdr.ts","./src/lib/ignore-pattern-sections.ts","./src/lib/library-file-column-widths.ts","./src/lib/library-file-search.ts","./src/lib/library-statistics-settings.test.ts","./src/lib/library-statistics-settings.ts","./src/lib/page-visibility.ts","./src/lib/paginated-files.test.ts","./src/lib/paginated-files.ts","./src/lib/path-display.test.ts","./src/lib/path-display.ts","./src/lib/resolution-categories.ts","./src/lib/scan-jobs.test.tsx","./src/lib/scan-jobs.tsx","./src/lib/settings-panel-state.ts","./src/lib/theme.tsx","./src/pages/dashboardpage.tsx","./src/pages/filedetailpage.test.tsx","./src/pages/filedetailpage.tsx","./src/pages/librariespage.test.tsx","./src/pages/librariespage.tsx","./src/pages/librarydetailpage.test.tsx","./src/pages/librarydetailpage.tsx","./src/test/setup.ts"],"version":"5.9.3"}
